pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "fastapi>=0.115.12",
    "openai>=1.82.1",
    "openai-agents>=0.0.16",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.5",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
